requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
